상품 상세 페이지 기능 구현 (#167)
* 상품 상세 페이지 관련 프로젝트 이동

* 상품 상세 페이지 기능 구현

---------

+++ 01_Backend/backend/src/main/java/org/example/backend/repository/shop/ProductImgRepository.java
@@ -23,11 +23,11 @@ import java.util.List;
  * 2024-04-30         sjuk2          최초 생성
  */
 @Repository
-public interface ProductImgRepository extends JpaRepository<ProductImage, Integer> {
+public interface ProductImgRepository extends JpaRepository<ProductImage, Long> {
     @Query(value = "SELECT PD_IMG_URL AS pdImgUrl\n" +
             "FROM TB_PRODUCT_IMAGE\n" +
             "WHERE PD_ID LIKE '%' || PD_ID || '%'\n" +
             "ORDER BY PD_IMG_ID"
     , nativeQuery = true)
-    List<IProductImgDto> findAllByPdIdContaining(@Param("pdId") Integer pdId);
+    List<IProductImgDto> findAllByPdIdContaining(@Param("pdId") Long pdId);
 }
